src/idm/pyproject.toml
[project]
name = "pokemon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.21.0",
    "dataclasses>=0.8",
    "einops>=0.8.1",
    "gputil>=1.4.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "psutil>=6.1.0",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "truss>=0.9.95",
    "vector-quantize-pytorch>=1.22.16",
    "wandb>=0.19.0",
    "yt-dlp>=2025.5.22",
]
